Execute componentDidMount/componentDidUpdate callbacks in same order as React
Previously the `componentDidUpdate` hook was called immediately after a
component was updated but `componentDidMount` was deferred until the end
of the render cycle in order to ensure that the corresponding DOM
element was actually present in the document when the callback was
invoked.

The result was that after an update to a component which mounted new
child component instances, the parent's `componentDidUpdate` would be
called before the child's `componentDidMount`. This broke the assumptions
of the React Dev Tools.

Fix the problem by following React's approach and enqueing both updates
and mount callbacks in a single queue which are invoked at the end of
the render cycle.